pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.19.0",
    "openai>=2.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "pymilvus"
]
